// src/app/types/statisticType.ts
export type StatisticType = {
    closedDurationAverage: number
    lifeTimeLost: number
    lifeTimePayments: number
    lifeTimeRequested: number
    lifeTimePaidAverage: number
    lifeTimeReceivedAverage: number
    openAmountAverage: number
    openDurationAverage: number
    closedContractsCount: number
    openContractsCount: number
    openContractAmount: number
    roi: number
    closedProjectedAmountAverage: number
    lifetimeProjectedAmountClosed: number
    closedRealAmountAverage: number
}